first test

+++ fabric-cli.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# 工具环境变量
+export FABRIC_CFG_PATH=${PWD}
+
 # 打印帮助
 function printHelp () {
   echo "使用说明："
@@ -64,6 +67,10 @@ function generateChannelArtifacts() {
     exit 1
   fi
 
+  if [ ! -d "channel-artifacts" ]; then
+    mkdir channel-artifacts
+  fi
+
   echo "##########################################################"
   echo "#########  Generating Orderer Genesis block ##############"
   echo "##########################################################"
@@ -106,39 +113,39 @@ function replacePrivateKey () {
     OPTS="-i"
   fi
 
-  cp docker-base/template.yaml docker-compose.yaml
+  cp template.yaml $COMPOSE_FILE
 
   CURRENT_DIR=$PWD
 
   cd crypto-config/peerOrganizations/bank1.samples.cn/ca/
   PRIV_KEY=$(ls *_sk)
   cd "$CURRENT_DIR"
-  sed $OPTS "s/CA_BANK1_PRIVATE_KEY/${PRIV_KEY}/g" docker-compose-e2e.yaml
+  sed $OPTS "s/CA_BANK1_PRIVATE_KEY/${PRIV_KEY}/g" $COMPOSE_FILE
 
   cd crypto-config/peerOrganizations/owner1.samples.cn/ca/
   PRIV_KEY=$(ls *_sk)
   cd "$CURRENT_DIR"
-  sed $OPTS "s/CA_OWNER1_PRIVATE_KEY/${PRIV_KEY}/g" docker-compose-e2e.yaml
+  sed $OPTS "s/CA_OWNER1_PRIVATE_KEY/${PRIV_KEY}/g" $COMPOSE_FILE
 
   cd crypto-config/peerOrganizations/storage1.samples.cn/ca/
   PRIV_KEY=$(ls *_sk)
   cd "$CURRENT_DIR"
-  sed $OPTS "s/CA_STORAGE1_PRIVATE_KEY/${PRIV_KEY}/g" docker-compose-e2e.yaml
+  sed $OPTS "s/CA_STORAGE1_PRIVATE_KEY/${PRIV_KEY}/g" $COMPOSE_FILE
 
   cd crypto-config/peerOrganizations/supervisor1.samples.cn/ca/
   PRIV_KEY=$(ls *_sk)
   cd "$CURRENT_DIR"
-  sed $OPTS "s/CA_SUPERVISOR1_PRIVATE_KEY/${PRIV_KEY}/g" docker-compose-e2e.yaml
+  sed $OPTS "s/CA_SUPERVISOR1_PRIVATE_KEY/${PRIV_KEY}/g" $COMPOSE_FILE
 
   # MaxOSX 特殊
   if [ "$ARCH" == "Darwin" ]; then
-    rm docker-compose-e2e.yamlt
+    rm -f ${COMPOSE_FILE}t
   fi
 }
 
 # 创建docker容器
 function networkCreate () {
-  CHANNEL_NAME=$CHANNEL_NAME TIMEOUT=$CLI_TIMEOUT DELAY=$CLI_DELAY docker-compose -f $COMPOSE_FILE create
+  CHANNEL_NAME=$CHANNEL_NAME TIMEOUT=$CLI_TIMEOUT DELAY=$CLI_DELAY docker-compose -f $COMPOSE_FILE up --no-start
   if [ $? -ne 0 ]; then
     echo "ERROR !!!! Unable to create network"
     exit 1
@@ -154,10 +161,19 @@ function networkStop () {
 }
 
 function networkClean () {
-    docker-compose -f $COMPOSE_FILE down
-    rm -fr channel-artifacts/
-    rm -fr crypto-config/
-    rm -f $COMPOSE_FILE
+    if [ -a "$COMPOSE_FILE" ]; then
+        docker-compose -f $COMPOSE_FILE down
+        rm -f $COMPOSE_FILE
+    fi
+
+    if [ -d "channel-artifacts" ]; then
+        rm -fr channel-artifacts
+    fi
+
+    if [ -d "crypto-config" ]; then
+        rm -fr crypto-config
+    fi
+
 }
 
 
@@ -209,6 +225,7 @@ confirm
 
 # 执行子命令
 if [ "${MODE}" == "create" ]; then
+  networkClean
   generateCerts
   generateChannelArtifacts
   replacePrivateKey
